Task için (!C)RUD operasyonları yazıldı.

+++ src/tasks/tasks.service.ts
@@ -15,12 +15,12 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) { }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ id, userId: user.id });
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -36,15 +36,19 @@ export class TasksService {
         return this.taskRepository.creteTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = this.taskRepository.delete({ id, userId: user.id });
         if ((await result).affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
 
         return await task.save();
